Fixed bugs

+++ gaus.h
@@ -4,7 +4,7 @@
 double *initMatrix(int,int);
 void fillMatrix(double *,int,int);
 void printMatrix(double *,int,int);
-double *getRow(int*,int,int,int);
+double *getRow(double *,int,int,int);
 void multiplyRow(double *,double,int);
 
 #endif
